Api/requirements.txt
Flask
pandas
google-cloud-storage
google-cloud-secret-manager
pymongo
gunicorn
